fix: tx need confirmation when using a custom network
