+++ packages/web3-providers/src/Web3/EVM/providers/BaseHosted.ts
@@ -75,7 +75,7 @@ export class BaseHostedProvider
     protected get options() {
         return {
             isSupportedAccount: () => true,
-            isSupportedChainId: () => true,
+            isSupportedChainId: (chainId: number) => chainId > 0 && Math.floor(chainId) === chainId,
             getDefaultAccount: () => '',
             getDefaultChainId,
             formatAddress: formatEthereumAddress,
